LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/Document-createElementNS-expected.txt


PASS createElementNS should use the global registry by default
PASS createElementNS should use the specified scoped registry
PASS createElementNS should create a builtin element regardless of a custom element registry specified
PASS createElementNS should use the specified global registry
PASS createElementNS should create an upgrade candidate when there is no matching definition in the specified registry
PASS document.createElement should create a builtin element with null registry if customElement is set to null
PASS document.createElement should create a custom element candidate with null registry if customElement is set to null
PASS document.createElement should create a defined custom element with null registry if customElement is set to null
PASS createElementNS should create an upgrade candidate and the candidate should be upgraded when the element is defined
PASS createElementNS on a non-HTML document should still handle registries correctly
